pythonServer/requirements.txt
fastapi
uvicorn
pytest
starlette
httpx
pytest-cov
python-dotenv